% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_generator.R
\name{key_generator}
\alias{key_generator}
\title{Generate unique keys for dataset}
\usage{
key_generator(n, key_length, ...)
}
\arguments{
\item{n}{Integer indicating the desired number of keys to generate.}

\item{key_length}{Integer indicating the desired length of they key strings.}

\item{...}{Additional arguments passed on to \code{stri_rand_strings} (e.g., a
\code{pattern} argument specifying character classes from which elements will be
drawn).}
}
\value{
Character vector of length \code{n} containing unique alphanumeric keys.
}
\description{
This function generates a unique set of alphanumeric keys that can be
assigned to a data.frame or other dataset. The code is adapted from that of
Julien Navarre in \href{https://stackoverflow.com/questions/29652356/generate-unique-alphanumeric-ids}{this Stack Overflow thread}.
Parameters match those of the \code{\link[stringi]{stri_rand_strings}} function.
}
\examples{
key_generator(n=10, key_length = 6)
key_generator(n=5, key_length = 11)
key_generator(n=15, key_length = 3)

## Add key to data.frame
df <- data.frame(x=letters[1:10], y = 1:10)
df <- cbind(
  data.frame(key = key_generator(n=nrow(df), key_length = 3)),
  df)
df
}
