% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packr.R
\docType{package}
\name{packr}
\alias{packr}
\alias{packr-package}
\title{packr: A package with a minimum example of package creation.}
\description{
This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/paezha/Reproducible-Research-Workflow}
}
\author{
Antonio Paez, School of Geography and Earth Sciences, McMaster University \email{paezha@mcmaster.ca}
}
