% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard error of a sample mean}
\usage{
se(x)
}
\arguments{
\item{x}{Vector of numbers}
}
\value{
Numeric value
}
\description{
Calculates the standard error of a sample mean. Note \code{NA} values are
automatically omitted by default.
}
\examples{
se(1:5)
se(c(1:3,NA,4:5))
}
