% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_base.R
\name{tmap_base}
\alias{tmap_base}
\title{Interactive map base plot}
\usage{
tmap_base(data = NULL)
}
\arguments{
\item{data}{Optional spatial data source containing the spatial information
to be plotted in the interactive map.}
}
\value{
A tmap object with \code{tmap_mode()} set to \code{view}.
}
\description{
This function provides a basic interactive map. By default it contains three
basemap layers: OpenStreet Maps streets, a global orthophoto, and a global
terrain map. This is indended to get things started with an interactive map
to which additional layers will be added by the user.
}
\examples{
library(tmap)
library(magrittr)

data('World', 'metro')

## Use the example world data to create an interactive map
tmap_base() +
  ## Add country polygons, colored by the Happy Planet Index
  tm_shape(World) +
  tm_polygons('HPI') +
  ## Add cities, sized proportional to population, with clustering
  tm_shape(metro) +
  tm_dots(
    size = 'pop2020',
    col = 'blue',
    clustering = TRUE)
}
