% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot_theme.R
\name{my_plot_theme}
\alias{my_plot_theme}
\title{My theme for a 7.5 by 5 inch plot}
\usage{
my_plot_theme()
}
\value{
themed ggplot
}
\description{
My theme for a 7.5 by 5 inch plot
}
\examples{
library(ggplot2)
library(magrittr)
library(dplyr)

x <-
  tibble(
    a = (1:10),
    b = seq(0.1, 1, 0.1)
  )

x \%>\%
  ggplot(
    aes(
      x = a,
      y = b)) +
  geom_point() +
  my_plot_theme()
}
