% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plot_theme.R
\name{my_plot_theme}
\alias{my_plot_theme}
\title{Plot theme for a faceted plot, purple and aqua}
\usage{
my_plot_theme()
}
\value{
themed ggplot
}
\description{
Plot theme for a faceted plot, purple and aqua
}
\examples{
library(ggplot2)
library(magrittr)
library(dplyr)

tibble(
  mishi = c(14, 2, 20),
  home = c(22, 4, 20)) \%>\%
  ggplot(aes(mishi, home)) +
  geom_point() +
  my_plot_theme()
}
