% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packr-package.R
\docType{package}
\name{packr-package}
\alias{packr}
\alias{packr-package}
\title{packr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).

This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/paezha/packr}
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
\keyword{internal}
