% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packr-package.R
\docType{package}
\name{packr-package}
\alias{packr}
\alias{packr-package}
\title{packr: A Minimal Example of Package Creation}
\description{
An example of a minimal package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/packr}
  \item \url{https://paezha.github.io/packr/}
  \item Report bugs at \url{https://github.com/paezha/packr/issues}
}

}
\author{
\strong{Maintainer}: Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

}
\keyword{internal}
