% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Title: Standard error of a sample mean}
\usage{
se(x)
}
\arguments{
\item{x:}{A vector of numbers}
}
\value{
nemeric value
}
\description{
Title: Standard error of a sample mean
}
\examples{
se(1:5)
}
